scout-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
apscheduler==3.10.4
slowapi==0.1.9
python-dotenv==1.0.1
